Refactored actual query building into RelevanceMatchQueryRewriter
