+++ x-pack/plugin/relevance-search/src/main/java/org/elasticsearch/xpack/relevancesearch/RelevanceSearchPlugin.java
@@ -35,7 +35,9 @@ import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.tracing.Tracer;
 import org.elasticsearch.watcher.ResourceWatcherService;
 import org.elasticsearch.xcontent.NamedXContentRegistry;
+import org.elasticsearch.xpack.relevancesearch.query.QueryFieldsResolver;
 import org.elasticsearch.xpack.relevancesearch.query.RelevanceMatchQueryBuilder;
+import org.elasticsearch.xpack.relevancesearch.query.RelevanceMatchQueryRewriter;
 import org.elasticsearch.xpack.relevancesearch.relevance.curations.CurationsService;
 import org.elasticsearch.xpack.relevancesearch.relevance.settings.RelevanceSettingsService;
 
@@ -52,8 +54,7 @@ public class RelevanceSearchPlugin extends Plugin implements ActionPlugin, Searc
         logger.info("Relevance Search Plugin loaded");
     }
 
-    private final SetOnce<RelevanceSettingsService> relevanceSettingsService = new SetOnce<>();
-    private final SetOnce<CurationsService> curationsService = new SetOnce<>();
+    private final SetOnce<RelevanceMatchQueryRewriter> relevanceMatchQueryRewriter = new SetOnce<>();
 
     @Override
     public List<RestHandler> getRestHandlers(
@@ -81,8 +82,8 @@ public class RelevanceSearchPlugin extends Plugin implements ActionPlugin, Searc
         return Collections.singletonList(
             new QuerySpec<QueryBuilder>(
                 RelevanceMatchQueryBuilder.NAME,
-                (in) -> new RelevanceMatchQueryBuilder(relevanceSettingsService.get(), curationsService.get(), in),
-                (parser) -> RelevanceMatchQueryBuilder.fromXContent(parser, relevanceSettingsService.get(), curationsService.get())
+                (in) -> new RelevanceMatchQueryBuilder(relevanceMatchQueryRewriter.get(), in),
+                (parser) -> RelevanceMatchQueryBuilder.fromXContent(parser, relevanceMatchQueryRewriter.get())
             )
         );
     }
@@ -103,9 +104,13 @@ public class RelevanceSearchPlugin extends Plugin implements ActionPlugin, Searc
         Tracer tracer,
         AllocationDeciders allocationDeciders
     ) {
-        relevanceSettingsService.set(new RelevanceSettingsService(client));
-        curationsService.set(new CurationsService(client));
 
-        return List.of(relevanceSettingsService.get(), curationsService.get());
+        RelevanceSettingsService relevanceSettingsService = new RelevanceSettingsService(client);
+        CurationsService curationsService = new CurationsService(client);
+        QueryFieldsResolver queryFieldsResolver = new QueryFieldsResolver();
+
+        relevanceMatchQueryRewriter.set(new RelevanceMatchQueryRewriter(relevanceSettingsService, curationsService, queryFieldsResolver));
+
+        return List.of(relevanceMatchQueryRewriter.get());
     }
 }
